docs: Adds readme to main directory
This adds a readme to the main directory for the project. There is a more detailed readme in the next folder down which I had to include as part of project requirements, but I noticed after committing that the project page in github did not have any description since that previous one was one directory down.